--- src/components/shared/confirm-modal/ConfirmModalMessage.ts
@@ -1,15 +1,0 @@
-import { confirmMessage } from "@/constants/message"
-
-export type MessageKey = keyof typeof ConfirmModalMessage
-
-const ConfirmModalMessage = {
-  uploadProfileImage: confirmMessage.editProfileImage,
-  resetProfileImage: confirmMessage.resetProfileImage,
-  deleteContent: confirmMessage.deleteContent,
-  cancleVote: confirmMessage.cancleVote,
-  reserveCoffeeChat: confirmMessage.reserveCoffeeChat,
-  deleteCoffeeChatBeforeDue: confirmMessage.deleteCoffeeChatBeforeDue,
-  deleteCoffeeChatAfterDue: confirmMessage.deleteCoffeeChatAfterDue,
-} as const
-
-export default ConfirmModalMessage
